Revert "book detail"
