+++ lib/screens/book-detail/book_detail.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class BookDetail extends StatefulWidget {
+  const BookDetail({Key? key}) : super(key: key);
+
+  @override
+  State<BookDetail> createState() => _BookDetailState();
+}
+
+class _BookDetailState extends State<BookDetail> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+          child: Text(
+        "Book detail",
+        style: TextStyle(fontSize: 40),
+      )),
+    );
+  }
+}
