+++ lib/screens/library/widgets/book_by_tags.dart
@@ -7,7 +7,7 @@ import 'package:library_online_application/api/book_api.dart';
 import 'package:library_online_application/icons/bee_app_icons.dart';
 import 'package:library_online_application/models/book.dart';
 import 'package:library_online_application/models/tag.dart';
-import 'package:library_online_application/screens/book-detail/book_detail_screen.dart';
+import 'package:library_online_application/screens/book-detail/book_detail.dart';
 
 class BookByTags extends StatefulWidget {
   final Tag? tag;
@@ -233,8 +233,7 @@ class _BookByTagsState extends State<BookByTags> {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                          builder: (context) =>
-                              BookDetail(bookId: books[_index].id.toString())),
+                          builder: (context) => const BookDetail()),
                     );
                   },
                   child: Text(
