+++ pubspec.yaml
@@ -45,8 +45,6 @@ dependencies:
   provider: ^6.0.3
   sliding_up_panel: ^2.0.0+1
   syncfusion_flutter_pdfviewer: ^20.1.52-beta
-  google_nav_bar: ^5.0.6
-  line_icons: ^2.0.1
 
 dev_dependencies:
   flutter_lints: ^2.0.1
